+++ Data/HttpGetClient.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public protocol HttpGetClient {
+    func get(from url: URL) -> Data?
+}
