--- Data/RemoteGetCookieMessage.swift
@@ -1,3 +1,0 @@
-
-import Foundation
-
