+++ Domain/GetCookie.swift
@@ -4,7 +4,7 @@ public protocol GetCookie {
     func get(getCookieModel: GetCookieModel, completion:  @escaping (Result<CookieModel, Error>) -> CookieModel)
 }
 
-public struct GetCookieModel: Codable {
+public struct GetCookieModel: Model {
     public var fortune: String
     
     public init(fortune: String = "") {
@@ -13,9 +13,6 @@ public struct GetCookieModel: Codable {
 }
 
 extension GetCookieModel {
-   public func toData() -> Data? {
-    return try? JSONEncoder().encode(self)
-    }
     
     public func convertFromData(_ data: Data?) -> GetCookieModel? {
         if let data = data {
