+++ Domain/Model.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+public protocol Model: Codable {
+    
+}
+
+public extension Model {
+    public func toData() -> Data? {
+        return try? JSONEncoder().encode(self)
+    }
+}
